move agent base classes to agent_base.py
